pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "medical-openenv"
version = "3.0.0"
description = "Synthetic healthcare OpenEnv for medical record cleaning and PHI redaction."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.0.0",
    "faker>=24.0.0",
    "openenv>=0.1.0,<1.0.0",
    "httpx>=0.27.0",
    "gradio>=4.0.0",
    "pandas>=2.0.0",
    "numpy>=1.26.0",
    "fastmcp>=0.1.0",
    "python-dotenv>=1.0.0",
    "openai>=1.14.0,<2.0.0",  # required by inference.py (spec-mandated entry point)
]

[project.optional-dependencies]
# Install with: pip install -e ".[gym]"
gym = ["gymnasium>=0.26.0"]

# Alias kept for backwards compatibility — openai is now a core dep above.
llm = ["openai>=1.14.0,<2.0.0"]

# Full ML stack for local dev / Colab (not installed in Docker).
# Install with: pip install -e ".[ml]"  or  pip install -r requirements-ml.txt
ml = [
    "torch>=2.2.0,<3.0.0",
    "transformers>=4.38.0,<5.0.0",
    "evaluate>=0.4.0,<1.0.0",
    "bert-score>=0.3.13,<1.0.0",
    "sentence-transformers>=2.7.0,<4.0.0",
]

# Install with: pip install -e ".[test]"
test = ["pytest>=8.0.0,<10.0.0", "hypothesis>=6.100.0"]

[project.scripts]
server = "src.main:main"

